refactor(step-3): extract Bear encoding helpers into src/operations/bear-encoding.ts
Moves the 4 Bear-format encoding helpers (decodeTagName, cleanBase64,
convertCoreDataTimestamp, convertDateToCoreDataTimestamp) out of utils.ts
into a dedicated operations/bear-encoding.ts. These share a single
concern — converting between Bear's on-disk/on-URL formats and JS types —
and are consumed by operations (notes, tags) plus the bear-add-file tool
via cleanBase64.

Co-locates the convertCoreDataTimestamp test into the new module's
bear-encoding.test.ts. Updates the notes.ts DECODED_TAG_TITLE sync
comment to reference the helper's new home.

All 37 unit tests still pass (1 moved from utils.test.ts to
bear-encoding.test.ts; utils.test.ts now has 18 of its previous 19
describe assertions remaining).
