+++ src/tags.ts
@@ -1,5 +1,5 @@
 import type { BearNote, BearTag } from './types.js';
-import { convertCoreDataTimestamp, decodeTagName } from './utils.js';
+import { convertCoreDataTimestamp, decodeTagName } from './operations/bear-encoding.js';
 import { logAndThrow, logger } from './logging.js';
 import { closeBearDatabase, openBearDatabase } from './database.js';
